Module amazon_sp_solicitations::apis::solicitations_api

Enums

CreateProductReviewAndSellerFeedbackSolicitationError

struct for typed errors of method create_product_review_and_seller_feedback_solicitation

GetSolicitationActionsForOrderError

struct for typed errors of method get_solicitation_actions_for_order

Functions

create_product_review_and_seller_feedback_solicitation

Sends a solicitation to a buyer asking for seller feedback and a product review for the specified order. Send only one productReviewAndSellerFeedback or free form proactive message per order. Usage Plan: | Rate (requests per second) | Burst | | –– | –– | | 1 | 5 | For more information, see "Usage Plans and Rate Limits" in the Selling Partner API documentation.

get_solicitation_actions_for_order

Returns a list of solicitation types that are available for an order that you specify. A solicitation type is represented by an actions object, which contains a path and query parameter(s). You can use the path and parameter(s) to call an operation that sends a solicitation. Currently only the productReviewAndSellerFeedbackSolicitation solicitation type is available. Usage Plan: | Rate (requests per second) | Burst | | –– | –– | | 1 | 5 | For more information, see "Usage Plans and Rate Limits" in the Selling Partner API documentation.